Factor out SLIM trainer
